fix: requested changes for renaming of blog-articles, pubDate and move MediumCardProps
